Reorganize file structure
- Move all third-party API helper functions to its own folder
- Refactor smaller components into its own files
