+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import './firebase';
+import './apis/firebase';
 import App from './App';
 import FirebaseLoginBarrier from './FirebaseLoginBarrier';
